+++ R/swimmerplot.R
@@ -8,9 +8,10 @@
 #' Join all tables from `.lookup$dataset` on `id` 
 #'
 #' @param .lookup the lookup table, loaded along with the database or result of [get_lookup()]
-#' @param id the patient identifier
+#' @param id the patient identifier. Will be coerced as numeric.
 #' @param group a grouping variable, given as "dataset$column"
 #' @param origin a variable to consider as time 0, given as "dataset$column"
+#' @param id_lim a numeric vector of length 2 providing the minimum and maximum `id` to subset on. 
 #' @param time_unit if `origin!=NULL`, the unit to measure time. One of `c("days", "weeks", "months", "years")`.
 #' @param aes_color either `variable` ("{dataset} - {column}") or `label` (the column label)
 #' @param plotly whether to use `{plotly}` to get an interactive plot
@@ -23,7 +24,7 @@
 #' #tm = read_trialmaster("filename.zip", pw="xx")
 #' tm = edc_example_plot()
 #' load_list(tm)
-#' p = edc_swimmerplot(.lookup)
+#' p = edc_swimmerplot(.lookup, id_lim=c(5,45))
 #' p2 = edc_swimmerplot(.lookup, origin="db0$date_naissance", time_unit="weeks")
 #' p3 = edc_swimmerplot(.lookup, group="db0$group", aes_color="label")
 #' \dontrun{
@@ -31,7 +32,7 @@
 #' htmlwidgets::saveWidget(p, "edc_swimmerplot.html", selfcontained=TRUE)
 #' }
 #' @importFrom cli cli_abort cli_warn
-#' @importFrom dplyr left_join mutate select slice
+#' @importFrom dplyr between left_join mutate select slice
 #' @importFrom forcats as_factor
 #' @importFrom ggplot2 aes facet_wrap geom_line geom_point ggplot labs
 #' @importFrom glue glue
@@ -43,6 +44,7 @@
 #' @importFrom tidyselect where
 edc_swimmerplot = function(.lookup=getOption("edc_lookup", NULL), ..., 
                        id="SUBJID", group=NULL, origin=NULL, 
+                       id_lim=NULL,
                        time_unit=c("days", "weeks", "months", "years"),
                        aes_color=c("variable", "label"), plotly=TRUE){
   check_dots_empty()
@@ -71,16 +73,33 @@ edc_swimmerplot = function(.lookup=getOption("edc_lookup", NULL), ...,
     cli_abort(c("None of the datasets contains a date column"))
   }
   
+  dbs = dbs %>% 
+    imap(~{
+      xid = suppressWarnings(as.numeric(.x$id))
+      pb = is.na(xid)!=is.na(.x$id)
+      if(any(pb)) cli_warn(c("NAs introduced by coercion to numeric in {.val {.y}${id}}. Is {.arg id} set correctly?", 
+                             i="Problematic value{?s}: {.val { .x$id[pb]}}"))
+      .x$id = xid
+      .x
+    })
+  
   dat = dbs %>% 
     imap(~{
       .x %>% 
         pivot_longer(-id) %>% 
-        mutate(label=unlist(var_label(.x)[name]) %||% name,
-               dataset=.y,
-               variable=paste0(toupper(dataset), " - ", toupper(name)))
+        mutate(
+          label=unlist(var_label(.x)[name]) %||% name,
+          dataset=.y,
+          variable=paste0(toupper(dataset), " - ", toupper(name))
+        )
     }) %>% 
     list_rbind() %>% 
     mutate(date=value)
+  
+  if(!is.null(id_lim)){
+    if(!is.numeric(id_lim) && length(id_lim)!=2) cli_abort("{.arg id_lim} should be a numeric vector of length 2")
+    dat = dat %>% filter(between(id, id_lim[1], id_lim[2]))
+  }
   
   if(!is.null(group)){
     dat_group = parse_var(group, id, parent)