+++ NEWS.md
@@ -3,6 +3,17 @@
 
 EDCimport is a package designed to easily import data from EDC software TrialMaster and Macro. Browse code at <https://github.com/DanChaltiel/EDCimport>.
 
+# EDCimport 0.4.0 <sub><sup>2023/xx/xx</sup></sub>
+
+#### New features
+
+- New argument `edc_swimmerplot(id_lim)` to subset the swimmer plot to some patients only.
+
+
+#### Bug fixes
+
+- 
+
 # EDCimport 0.3.0 <sub><sup>2023/05/19</sup></sub>
 
 #### New features